fix endpoint management

+++ app/Admin/Controller/Endpoint/Delete.php
@@ -7,8 +7,11 @@ namespace KejawenLab\Application\Admin\Controller\Endpoint;
 use KejawenLab\ApiSkeleton\Security\Annotation\Permission;
 use KejawenLab\Application\Node\EndpointService;
 use KejawenLab\Application\Node\Model\EndpointInterface;
+use KejawenLab\Application\Node\Model\NodeInterface;
+use KejawenLab\Application\Node\NodeService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
@@ -19,18 +22,26 @@ use Symfony\Component\Routing\Annotation\Route;
  */
 final class Delete extends AbstractController
 {
-    public function __construct(private EndpointService $service)
+    public function __construct(private EndpointService $service, private NodeService $nodeService)
     {
     }
 
-    #[Route(path: '/services/endpoints/{id}/delete', name: Delete::class, methods: ['GET'])]
-    public function __invoke(string $id): Response
+    #[Route(path: '/services/nodes/{nodeId}/endpoints/{id}/delete', name: Delete::class, methods: ['GET'])]
+    public function __invoke(Request $request, string $nodeId, string $id): Response
     {
+        /** @var NodeInterface $node */
+        $node = $this->nodeService->get($nodeId);
+        if (null === $node) {
+            $this->addFlash('error', 'sas.page.node.not_found');
+
+            return new RedirectResponse($this->generateUrl(Main::class, $request->query->all()));
+        }
+
         $endpoint = $this->service->get($id);
         if (!$endpoint instanceof EndpointInterface) {
             $this->addFlash('error', 'sas.page.endpoint.not_found');
 
-            return new RedirectResponse($this->generateUrl(Main::class));
+            return new RedirectResponse($this->generateUrl(Main::class, $request->query->all()));
         }
 
         $this->service->delete($endpoint);
